pyproject.toml
[project]
name = "mcp-jira"
version = "0.1.0"
description = "Async MCP Server for JIRA integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logging>=0.4.9.6",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
    "typing>=3.10.0.0",
    "aiohttp>=3.9.0",
    "asyncio-throttle>=1.0.2",
]
